Moved OpenSSL managed ptrs to better place
